Refactor model architecture: Split monolithic model.py by generation
- Split model/model.py into model/qwen2_5_vl.py and model/qwen3.py for better architectural separation
- Created separate config classes (Qwen2Config, Qwen3Config) for each generation
- Implemented dynamic config loading via get_config_class() methods in each model class
- Updated imports in run.py to use the new split files
- Modified util.py to use dynamic config resolution instead of shared ModelConfig
- Maintains universal loading functionality while providing clear architectural boundaries
- Supports future model generations (Qwen4, Qwen3-VL) without conditional complexity
